fbxをオリジナルフォーマットに

+++ lib/Engine/2D/PostEffect/GaussianBlurPostEffect.cpp
@@ -33,14 +33,14 @@ void GaussianBlurPostEffect::PostUpdate(RenderTarget* mainRenderTarget)
 void GaussianBlurPostEffect::UpdateWeightsTable(float blurPower)
 {
 	float total = 0;
-	for (int i = 0; i < NUM_WEIGHTS; i++)
+	for (size_t i = 0; i < NUM_WEIGHTS; i++)
 	{
 		weight[i] = expf(-0.5f * (float)(i * i) / blurPower);
 		total += 2.0f * weight[i];
 
 	}
 	// ‹KŠi‰»
-	for (int i = 0; i < NUM_WEIGHTS; i++)
+	for (size_t i = 0; i < NUM_WEIGHTS; i++)
 	{
 		weight[i] /= total;
 	}
